# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Réservation",
        "description": "API for booking properties and apartments",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/owner/properties/{property}/apartments/{apartment}/bookings": {
            "get": {
                "tags": [
                    "Owner Apartments"
                ],
                "summary": "Get all bookings for a specific apartment",
                "operationId": "d575a80345a07113be5effc0c5877c4c",
                "parameters": [
                    {
                        "name": "property",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apartment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/owner/properties/{property}/apartments": {
            "get": {
                "tags": [
                    "Owner Apartments"
                ],
                "summary": "Get all apartments for a property",
                "operationId": "93cc67ce66eb297b09611adc705a420c",
                "parameters": [
                    {
                        "name": "property",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ApartmentDetailsResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Owner Apartments"
                ],
                "summary": "Create a new apartment",
                "operationId": "31d9506733a6feb0da959f5f23069763",
                "parameters": [
                    {
                        "name": "property",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "apartment_type_id",
                                    "size"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Apartment 101"
                                    },
                                    "apartment_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "size": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Apartment created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/owner/properties/{property}/apartments/{apartment}": {
            "get": {
                "tags": [
                    "Owner Apartments"
                ],
                "summary": "Get apartment details",
                "operationId": "d64544ee09d4dad29bc184718780f25a",
                "parameters": [
                    {
                        "name": "property",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apartment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApartmentDetailsResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Owner Apartments"
                ],
                "summary": "Update an apartment",
                "operationId": "e7b9dd1daa64fa3208a51b04bf6293c2",
                "parameters": [
                    {
                        "name": "property",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apartment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Apartment"
                                    },
                                    "size": {
                                        "type": "integer",
                                        "example": 60
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Apartment updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "updated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/owner/properties/{property}/apartments/{apartment}/deactivate": {
            "put": {
                "tags": [
                    "Owner Apartments"
                ],
                "summary": "Deactivate an apartment",
                "operationId": "38027a73567b6d47235809e870d5a936",
                "parameters": [
                    {
                        "name": "property",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apartment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Apartment deactivated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "deactivated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/owner/properties/{property}/apartments/{apartment}/activate": {
            "put": {
                "tags": [
                    "Owner Apartments"
                ],
                "summary": "Activate an apartment",
                "operationId": "f820db10f09786d4a579f0de21155fbd",
                "parameters": [
                    {
                        "name": "property",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apartment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Apartment activated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "activated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApartmentDetailsResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Apartment 101"
                    },
                    "active": {
                        "type": "string",
                        "example": "yes"
                    },
                    "type": {
                        "type": "string",
                        "example": "Studio"
                    },
                    "adult_capacity": {
                        "type": "integer",
                        "example": 2
                    },
                    "children_capacity": {
                        "type": "integer",
                        "example": 1
                    },
                    "size": {
                        "type": "integer",
                        "example": 50
                    },
                    "beds_list": {
                        "type": "string",
                        "example": "1 double bed"
                    },
                    "bathrooms": {
                        "type": "integer",
                        "example": 1
                    },
                    "facilities": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "available_in": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "booked_in": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "BookingResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "apartment_name": {
                        "type": "string",
                        "example": "Apartment 101"
                    },
                    "apartment_view_link": {
                        "type": "string",
                        "example": "http://example.com/api/v1/apartments/view/1"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-03-15"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-03-20"
                    },
                    "adult_guests": {
                        "type": "integer",
                        "example": 2
                    },
                    "children_guests": {
                        "type": "integer",
                        "example": 1
                    },
                    "total_price": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "rating": {
                        "type": "integer",
                        "example": 9
                    },
                    "review_comment": {
                        "type": "string",
                        "example": "Great stay!"
                    },
                    "booked_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-03-01"
                    },
                    "cancelled_at": {
                        "type": "string",
                        "format": "date",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}